+++ geolocator_platform_interface/test/geolocator_platform_interface_test.dart
@@ -82,7 +82,8 @@ void main() {
 
       // Act & Assert
       expect(
-        geolocatorPlatform.requestTemporaryFullAccuracy,
+        geolocatorPlatform.requestTemporaryFullAccuracy(
+            purposeKey: 'purposeKey'),
         throwsUnimplementedError,
       );
     });